backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.1
pydantic>=2.7.0
langchain>=0.1.9
langchain-google-genai>=0.0.9
langchain-community>=0.0.24
python-dotenv>=1.0.1
tavily-python>=0.3.1   